Rewrite MapzenMap as RegionMap
 - Changes most options to public fields
 - Removes 'Region Scale Units' in favor of a simple 'Units per Meter'
 - Refactors and simplifies parts of SceneGroup and SceneGraph
 - Moves SceneGroup.Type enum to SceneGroupType
 - Rewrites MapzenMapEditor as RegionMapEditor with improvements
