+++ Assets/Mapzen/Unity/Editor/RegionMapEditor.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 87939f0ef9919450f85d1ec5b75ca879
-timeCreated: 1506545153
+guid: 928281946bbb749c4895d359cca52332
+timeCreated: 1510952701
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
